debt - split widget and controller into separate files (#179237)
